Factorize waiting for channel ready handlers
